Integrated generating student trends in weekly scheduled firebase function - Moved code to update trends collections to server side - Deployed function
