+++ functions/package.json
@@ -17,11 +17,15 @@
   "main": "lib/index.js",
   "dependencies": {
     "@genkit-ai/google-genai": "^1.27.0",
+    "@tensorflow-models/universal-sentence-encoder": "^1.3.3",
+    "@tensorflow/tfjs": "^4.22.0",
+    "@xenova/transformers": "^2.17.2",
     "firebase": "^11.2.0",
     "firebase-admin": "^13.0.2",
     "firebase-functions": "^6.3.2",
     "firebase-tools": "^13.6.0",
     "genkit": "^1.27.0",
+    "kmeans-ts": "^1.0.4",
     "moment-timezone": "^0.5.45",
     "twilio": "^3.71.3"
   },
